feat: expira sessoes inativas e limita tentativas de entrada por codigo
Um BackgroundService varre as sessoes a cada minuto e descarta as que passaram
do tempo de inatividade, avisando com SessionEnded quem ainda estiver com a
pagina aberta. A janela e deslizante: conversar mantem a sessao viva.

O limitador nativo entra por politica em cada endpoint, e nao globalmente, para
nao tocar no hub: 20 consultas de codigo e 10 criacoes por minuto por IP. E o
que sustenta o codigo de 6 caracteres, curto o bastante para ser digitado.

SessionTimeoutMinutes passa a aceitar fracao de minuto, o que permite exercitar
expiracao e janela deslizante sem esperar meia hora.

+++ Options/ShareLinkOptions.cs
@@ -8,7 +8,7 @@ public sealed class ShareLinkOptions
     public const string SectionName = "ShareLink";
 
     /// <summary>Tempo de inatividade após o qual a sessão é descartada. Janela deslizante: renova a cada mensagem ou entrada.</summary>
-    public int SessionTimeoutMinutes { get; set; } = 30;
+    public double SessionTimeoutMinutes { get; set; } = 30;
 
     /// <summary>Quantas mensagens recentes a sessão guarda para reenviar a quem entra ou reconecta.</summary>
     public int MaxMessagesPerSession { get; set; } = 50;
